fix(lint): split ui/routines/page.rs under the 600-line gate
