+++ ui/src/routines/mod.rs
@@ -9,11 +9,14 @@
 //! - `state` — page/view/modal state, group-by, sorting, and the reducer.
 //! - `hooks` — small custom Yew hooks used by the list page.
 //! - `page` — the `RoutinesPage` component that wires the above together.
+//! - `bulk_actions` — the list page's bulk-selection callbacks (select/select-all,
+//!   bulk enable/disable/delete).
 //! - `banner`, `filter_bar`, `calendar`, `table`, `row`, `form`, `bulk`, `logs`,
 //!   `flags_panel` — the list page's sub-components.
 
 mod banner;
 mod bulk;
+mod bulk_actions;
 mod calendar;
 mod filter;
 mod filter_bar;
